Refactor constants into data directory
