--- src/public/constants/mousePhenotypesTopLevel.js
@@ -1,509 +1,0 @@
-import _ from 'lodash';
-
-// https://www.ebi.ac.uk/ols//api/ontologies/mp/terms/http%253A%252F%252Fpurl.obolibrary.org%252Fobo%252FMP_0000001/jstree/children/99815577_1
-const mousePhenotypesTopLevel = [
-  {
-    id: '103407902_child_1',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0002006',
-    text: 'neoplasm',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0002006',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0002006',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '97042208_child_2',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0010771',
-    text: 'integument phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0010771',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0010771',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99055684_child_3',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0010768',
-    text: 'mortality/aging',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0010768',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0010768',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99828977_child_4',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0002873',
-    text: 'normal phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0002873',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0002873',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99239513_child_5',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0001186',
-    text: 'pigmentation phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0001186',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0001186',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99832633_child_6',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0003631',
-    text: 'nervous system phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0003631',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0003631',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99826779_child_7',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0003012',
-    text: 'no phenotypic analysis',
-    state: {
-      opened: true,
-    },
-    children: false,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0003012',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0003012',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99248775_child_8',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005367',
-    text: 'renal/urinary system phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005367',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005367',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99819782_child_9',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005369',
-    text: 'muscle phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005369',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005369',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99819740_child_10',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005370',
-    text: 'liver/biliary system phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005370',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005370',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99819737_child_11',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005371',
-    text: 'limbs/digits/tail phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005371',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005371',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '94020203_child_12',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005377',
-    text: 'hearing/vestibular/ear phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005377',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005377',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '94012547_child_13',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005378',
-    text: 'growth/size/body region phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005378',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005378',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99815788_child_14',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005375',
-    text: 'adipose tissue phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005375',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005375',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99061541_child_15',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005376',
-    text: 'homeostasis/metabolism phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005376',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005376',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99815782_child_16',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005379',
-    text: 'endocrine/exocrine gland phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005379',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005379',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99055380_child_17',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005380',
-    text: 'embryo phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005380',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005380',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99055595_child_18',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005381',
-    text: 'digestive/alimentary phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005381',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005381',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '94734251_child_19',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005384',
-    text: 'cellular phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005384',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005384',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '93896082_child_20',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005385',
-    text: 'cardiovascular system phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005385',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005385',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99815721_child_21',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005382',
-    text: 'craniofacial phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005382',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005382',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '94730539_child_22',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005388',
-    text: 'respiratory system phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005388',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005388',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '94725436_child_23',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005389',
-    text: 'reproductive system phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005389',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005389',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99815712_child_24',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005386',
-    text: 'behavior/neurological phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005386',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005386',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99060410_child_25',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005387',
-    text: 'immune system phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005387',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005387',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '99815630_child_26',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005391',
-    text: 'vision/eye phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005391',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005391',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '93897320_child_27',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005390',
-    text: 'skeleton phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005390',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005390',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '97041369_child_28',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005394',
-    text: 'taste/olfaction phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005394',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005394',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-  {
-    id: '93295241_child_29',
-    parent: '99815577_1',
-    iri: 'http://purl.obolibrary.org/obo/MP_0005397',
-    text: 'hematopoietic system phenotype',
-    state: {
-      opened: false,
-    },
-    children: true,
-    a_attr: {
-      iri: 'http://purl.obolibrary.org/obo/MP_0005397',
-      ontology_name: 'mp',
-      title: 'http://purl.obolibrary.org/obo/MP_0005397',
-      class: 'is_a',
-    },
-    ontology_name: 'mp',
-  },
-];
-
-export default _.sortBy(
-  mousePhenotypesTopLevel.map(d => ({
-    id: d.iri
-      .split('/')
-      .pop()
-      .replace('_', ':'),
-    name: d.text.replace(' phenotype', ''),
-  })),
-  'name'
-);
