--- src/public/constants/omnipathCategories.js
@@ -1,22 +1,0 @@
-export default {
-  Pathways: {
-    SignaLink3: true,
-    Signor: true,
-    Reactome: true,
-    SPIKE: true,
-  },
-  EnzymeSubstrate: {
-    PhosphoPoint: true,
-    HPRD: true,
-    'HPRD-phos': true,
-    MIMP: true,
-    HuPho: true,
-  },
-  PPI: {
-    BioGRID: true,
-    InnateDB: true,
-    IntAct: true,
-    DIP: true,
-    STRING: true,
-  },
-};
