--- src/public/constants/reactomeTopLevel.js
@@ -1,401 +1,0 @@
-// https://reactome.org/ContentService/data/pathways/top/9606
-const reactomeTopLevel = [
-  {
-    dbId: 1640170,
-    displayName: 'Cell Cycle',
-    stId: 'R-HSA-1640170',
-    stIdVersion: 'R-HSA-1640170.2',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Cell Cycle'],
-    releaseDate: '2011-12-06',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 1500931,
-    displayName: 'Cell-Cell communication',
-    stId: 'R-HSA-1500931',
-    stIdVersion: 'R-HSA-1500931.3',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Cell-Cell communication'],
-    releaseDate: '2011-09-20',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 8953897,
-    displayName: 'Cellular responses to external stimuli',
-    stId: 'R-HSA-8953897',
-    stIdVersion: 'R-HSA-8953897.2',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Cellular responses to external stimuli'],
-    releaseDate: '2017-03-27',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 4839726,
-    displayName: 'Chromatin organization',
-    stId: 'R-HSA-4839726',
-    stIdVersion: 'R-HSA-4839726.2',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Chromatin organization'],
-    releaseDate: '2013-12-04',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 400253,
-    displayName: 'Circadian Clock',
-    stId: 'R-HSA-400253',
-    stIdVersion: 'R-HSA-400253.2',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Circadian Clock'],
-    releaseDate: '2010-12-14',
-    speciesName: 'Homo sapiens',
-    doi: '10.3180/REACT_24941.1',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 1266738,
-    displayName: 'Developmental Biology',
-    stId: 'R-HSA-1266738',
-    stIdVersion: 'R-HSA-1266738.4',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Developmental Biology'],
-    releaseDate: '2011-09-20',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 8963743,
-    displayName: 'Digestion and absorption',
-    stId: 'R-HSA-8963743',
-    stIdVersion: 'R-HSA-8963743.1',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Digestion and absorption'],
-    releaseDate: '2017-03-27',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 1643685,
-    displayName: 'Disease',
-    stId: 'R-HSA-1643685',
-    stIdVersion: 'R-HSA-1643685.6',
-    isInDisease: true,
-    isInferred: false,
-    name: ['Disease'],
-    releaseDate: '2011-12-06',
-    releaseStatus: 'UPDATED',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 73894,
-    displayName: 'DNA Repair',
-    stId: 'R-HSA-73894',
-    stIdVersion: 'R-HSA-73894.2',
-    isInDisease: false,
-    isInferred: false,
-    name: ['DNA Repair'],
-    releaseDate: '2004-07-06',
-    releaseStatus: 'UPDATED',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 69306,
-    displayName: 'DNA Replication',
-    stId: 'R-HSA-69306',
-    stIdVersion: 'R-HSA-69306.3',
-    isInDisease: false,
-    isInferred: false,
-    name: ['DNA Replication'],
-    releaseDate: '2003-01-06',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 1474244,
-    displayName: 'Extracellular matrix organization',
-    stId: 'R-HSA-1474244',
-    stIdVersion: 'R-HSA-1474244.2',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Extracellular matrix organization'],
-    releaseDate: '2012-03-13',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 74160,
-    displayName: 'Gene expression (Transcription)',
-    stId: 'R-HSA-74160',
-    stIdVersion: 'R-HSA-74160.6',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Gene expression (Transcription)'],
-    releaseDate: '2004-07-06',
-    releaseStatus: 'UPDATED',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 109582,
-    displayName: 'Hemostasis',
-    stId: 'R-HSA-109582',
-    stIdVersion: 'R-HSA-109582.1',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Hemostasis', 'Blood coagulation'],
-    releaseDate: '2005-04-11',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 168256,
-    displayName: 'Immune System',
-    stId: 'R-HSA-168256',
-    stIdVersion: 'R-HSA-168256.6',
-    isInDisease: false,
-    isInferred: false,
-    name: [
-      'Immune System',
-      'Signaling in Immune system',
-      'Immune System signaling',
-    ],
-    releaseDate: '2006-04-24',
-    releaseStatus: 'UPDATED',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 1430728,
-    displayName: 'Metabolism',
-    stId: 'R-HSA-1430728',
-    stIdVersion: 'R-HSA-1430728.8',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Metabolism'],
-    releaseDate: '2011-09-20',
-    releaseStatus: 'UPDATED',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 392499,
-    displayName: 'Metabolism of proteins',
-    stId: 'R-HSA-392499',
-    stIdVersion: 'R-HSA-392499.6',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Metabolism of proteins'],
-    releaseDate: '2009-04-01',
-    releaseStatus: 'UPDATED',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 8953854,
-    displayName: 'Metabolism of RNA',
-    stId: 'R-HSA-8953854',
-    stIdVersion: 'R-HSA-8953854.3',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Metabolism of RNA'],
-    releaseDate: '2004-07-06',
-    speciesName: 'Homo sapiens',
-    doi: '10.3180/R-HSA-8953854.1',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 5205647,
-    displayName: 'Mitophagy',
-    stId: 'R-HSA-5205647',
-    stIdVersion: 'R-HSA-5205647.2',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Mitophagy'],
-    releaseDate: '2015-09-22',
-    speciesName: 'Homo sapiens',
-    doi: '10.3180/r-hsa-5205647',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 397014,
-    displayName: 'Muscle contraction',
-    stId: 'R-HSA-397014',
-    stIdVersion: 'R-HSA-397014.2',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Muscle contraction'],
-    releaseDate: '2009-04-01',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 112316,
-    displayName: 'Neuronal System',
-    stId: 'R-HSA-112316',
-    stIdVersion: 'R-HSA-112316.5',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Neuronal System'],
-    releaseDate: '2008-06-30',
-    releaseStatus: 'UPDATED',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 1852241,
-    displayName: 'Organelle biogenesis and maintenance',
-    stId: 'R-HSA-1852241',
-    stIdVersion: 'R-HSA-1852241.4',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Organelle biogenesis and maintenance'],
-    releaseDate: '2014-03-12',
-    releaseStatus: 'UPDATED',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 5357801,
-    displayName: 'Programmed Cell Death',
-    stId: 'R-HSA-5357801',
-    stIdVersion: 'R-HSA-5357801.1',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Programmed Cell Death'],
-    releaseDate: '2014-12-11',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 9609507,
-    displayName: 'Protein localization',
-    stId: 'R-HSA-9609507',
-    stIdVersion: 'R-HSA-9609507.1',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Protein localization'],
-    releaseDate: '2018-09-12',
-    releaseStatus: 'NEW',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 1474165,
-    displayName: 'Reproduction',
-    stId: 'R-HSA-1474165',
-    stIdVersion: 'R-HSA-1474165.2',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Reproduction'],
-    releaseDate: '2013-06-11',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 162582,
-    displayName: 'Signal Transduction',
-    stId: 'R-HSA-162582',
-    stIdVersion: 'R-HSA-162582.7',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Signal Transduction', 'Signaling Pathways'],
-    releaseDate: '2011-09-20',
-    releaseStatus: 'UPDATED',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 382551,
-    displayName: 'Transport of small molecules',
-    stId: 'R-HSA-382551',
-    stIdVersion: 'R-HSA-382551.4',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Transport of small molecules'],
-    releaseDate: '2008-12-17',
-    speciesName: 'Homo sapiens',
-    doi: '10.3180/REACT_15518.1',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-  {
-    dbId: 5653656,
-    displayName: 'Vesicle-mediated transport',
-    stId: 'R-HSA-5653656',
-    stIdVersion: 'R-HSA-5653656.1',
-    isInDisease: false,
-    isInferred: false,
-    name: ['Vesicle-mediated transport'],
-    releaseDate: '2015-03-19',
-    speciesName: 'Homo sapiens',
-    hasDiagram: true,
-    schemaClass: 'TopLevelPathway',
-    className: 'TopLevelPathway',
-  },
-];
-
-export default reactomeTopLevel.map(d => ({ id: d.stId, name: d.displayName }));
