+++ src/public/data/omnipath/interactionCategories.json
@@ -1,0 +1,22 @@
+{
+  "Pathways": {
+    "SignaLink3": true,
+    "Signor": true,
+    "Reactome": true,
+    "SPIKE": true
+  },
+  "EnzymeSubstrate": {
+    "PhosphoPoint": true,
+    "HPRD": true,
+    "HPRD-phos": true,
+    "MIMP": true,
+    "HuPho": true
+  },
+  "PPI": {
+    "BioGRID": true,
+    "InnateDB": true,
+    "IntAct": true,
+    "DIP": true,
+    "STRING": true
+  }
+}
