+++ src/public/data/uniprot/subcellularLocations.json
