+++ src/public/data/uniprot/uniprotIdToGene.json
